data-selection/requirements.txt
ray>=2.9.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
datasets>=2.18.0
Pillow>=10.0.0
numpy>=1.26.0
